Fixed #1 - SeqIO does not correctly convert gzipped files.
Changed how we handle sequence files. If the file ends in a .gz
extension, then use gzip.open() as the file handle, otherwise use a
regular file handle. If SeqIO raises an exception (because the file is
already in sanger fastq format) then either copy the file or gzip it
directly.

 Please enter the commit message for your changes. Lines starting